+++ fsl/wrappers/flirt.py
@@ -4,6 +4,7 @@
 #
 # Author: Sean Fitzgibbon <sean.fitzgibbon@ndcn.ox.ac.uk>
 # Author: Paul McCarthy <pauldmccarthy@gmail.com>
+# Author: Fidel Alfaro Almagro <fidel.alfaroalmagro@ndcn.ox.ac.uk>
 #
 """This module provides wrapper functions for the FSL `FLIRT
 <https://fsl.fmrib.ox.ac.uk/fsl/fslwiki/FLIRT>`_ tool, and other related
@@ -17,7 +18,9 @@ tools.
    applyxfm4D
    invxfm
    concatxfm
+   fixscaleskew
    mcflirt
+   standard_space_roi
 """
 
 
@@ -125,6 +128,30 @@ def concatxfm(atob, btoc, atoc):
     return cmd
 
 
+@wutils.fileOrArray('inmat1', 'inmat2', 'omat')
+@wutils.fslwrapper
+def fixscaleskew(inmat1, inmat2, omat):
+    """Use ``convert_xfm`` to  fix the skew scale of an affine. Note that
+    the order of the input matrices is the opposite of the order expected
+    by ``convert_xfm``.
+
+    :arg inmat1: Input matrix to fix.
+    :arg inmat2: Matrix to scale the skew to.
+    :arg omat:   Output matrix.
+    """
+
+    asrt.assertFileExists(inmat1, inmat2)
+
+    cmd = ['convert_xfm',
+           '-fixscaleskew',
+           inmat2,
+           inmat1,
+           '-omat',
+           omat]
+
+    return cmd
+
+
 @wutils.fileOrImage('infile', 'out', 'reffile', outprefix='out')
 @wutils.fileOrArray('init', outprefix='out')
 @wutils.fslwrapper
@@ -148,6 +175,46 @@ def mcflirt(infile, **kwargs):
     }
 
     cmd  = ['mcflirt', '-in', infile]
+    cmd += wutils.applyArgStyle('-', argmap=argmap, valmap=valmap, **kwargs)
+
+    return cmd
+
+
+@wutils.fileOrImage('input', 'output', 'maskMASK', 'roiMASK',
+                    'ssref', 'altinput')
+@wutils.fslwrapper
+def standard_space_roi(input, output, **kwargs):
+    """Wrapper for the ``standard_space_roi`` command.
+
+    Refer to the ``standard_space_roi`` command-line help for details on all
+    arguments.
+    """
+    asrt.assertIsNifti(input)
+
+    argmap = {
+        'twod' : '2D'
+    }
+
+    valmap = {
+        'maskFOV'      : wutils.SHOW_IF_TRUE,
+        'maskNONE'     : wutils.SHOW_IF_TRUE,
+        'roiFOV'       : wutils.SHOW_IF_TRUE,
+        'roiNONE'      : wutils.SHOW_IF_TRUE,
+        'd'            : wutils.SHOW_IF_TRUE,
+        'b'            : wutils.SHOW_IF_TRUE,
+        'usesqform'    : wutils.SHOW_IF_TRUE,
+        'displayinit'  : wutils.SHOW_IF_TRUE,
+        'noresample'   : wutils.SHOW_IF_TRUE,
+        'forcescaling' : wutils.SHOW_IF_TRUE,
+        'applyxfm'     : wutils.SHOW_IF_TRUE,
+        'nosearch'     : wutils.SHOW_IF_TRUE,
+        'noclamp'      : wutils.SHOW_IF_TRUE,
+        'noresampblur' : wutils.SHOW_IF_TRUE,
+        '2D'           : wutils.SHOW_IF_TRUE,
+        'v'            : wutils.SHOW_IF_TRUE
+    }
+
+    cmd  = ['standard_space_roi', input, output]
     cmd += wutils.applyArgStyle('-', argmap=argmap, valmap=valmap, **kwargs)
 
     return cmd