+++ CHANGELOG.rst
@@ -13,6 +13,8 @@ Added
 * New :func:`.standard_space_roi` and :func:`.fslswapdim` wrapper functions
   for the FSL commands (!351, !354).
 * New :mod:`fsl.wrappers.first` wrapper functions (!355).
+* New :mod:`fsl.wrappers.bianca` and :mod:`fsl.wrappers.avwutils` wrapper
+  functions (!358).
 * New :func:`.wrapperconfig` context manager function, which allows the
   default values for arguments passed by the :mod:`fsl.wrappers` functions to
   the :func:`fsl.utils.run.run` function to be changed (!352, !356).